Smartcar Changes
Trying to make a class that with one call will have the client authenticate (because they have to do so) and then return the odometer and vehicle info. I also messed with the server.py class a bit,  I can not test it however because I am having massive flask issues.